+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/token/XsuaaToken.java
@@ -1,5 +1,6 @@
 package com.sap.cloud.security.xsuaa.token;
 
+import com.sap.cloud.security.token.InvalidTokenException;
 import net.minidev.json.JSONArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,6 +13,8 @@ import org.springframework.util.Assert;
 import java.time.Instant;
 import java.util.*;
 
+import static com.sap.cloud.security.token.TokenClaims.AUTHORIZATION_PARTY;
+import static com.sap.cloud.security.token.TokenClaims.XSUAA.CLIENT_ID;
 import static com.sap.cloud.security.xsuaa.token.TokenClaims.*;
 import static org.springframework.util.StringUtils.isEmpty;
 
@@ -141,7 +144,24 @@ public class XsuaaToken extends Jwt implements Token {
 	@Override
 	@Nullable
 	public String getClientId() {
-		return getClaimAsString(CLAIM_CLIENT_ID);
+		String clientId = getClaimAsString(AUTHORIZATION_PARTY);
+		if (clientId == null || clientId.trim().isEmpty()) {
+			List<String> audiences = getAudience();
+
+			if (audiences != null && audiences.size() == 1) {
+				return audiences.get(0);
+			} else if (hasClaim(CLIENT_ID) && !getClaimAsString(CLIENT_ID).trim() // required for backward compatibility
+																					// for generated tokens in JUnit
+																					// tests
+					.isEmpty()) {
+				logger.warn("usage of 'cid' claim is deprecated and should be replaced by 'azp' or 'aud' claims");
+				return getClaimAsString(CLIENT_ID);
+			}
+			logger.error("Couldn't get client id. Invalid authorized party or audience claims.");
+			throw new InvalidTokenException("Couldn't get client id. Invalid authorized party or audience claims.");
+		} else {
+			return clientId;
+		}
 	}
 
 	@Override
@@ -224,7 +244,7 @@ public class XsuaaToken extends Jwt implements Token {
 
 	/**
 	 * Check if the authentication token contains a claim, e.g. "email".
-	 * 
+	 *
 	 * @param claim
 	 *            name of the claim
 	 * @return true: attribute exists